+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Add missing type hint in table filter extension
+
 ## [2.0.2] - 2020-07-28
 [2.0.2]: https://github.com/Smile-SA/gdpr-dump/compare/2.0.1...2.0.2